@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { UserInfoService } from './user-info.service';
-import 'rxjs/add/observable/of';
+
 
 @Injectable()
 export class AuthGuard implements CanActivate {
