@@ -2,9 +2,9 @@ import { Injectable } from '@angular/core';
 
 import { Resolve } from '@angular/router';
 
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/delay';
+import { Observable } from 'rxjs';
+
+
 import { UserInfoService } from '../../core/user-info.service';
 
 @Injectable()
