@@ -2,7 +2,7 @@ import { Component, OnInit, OnDestroy, Renderer2, ViewChild, ElementRef } from '
 import { ShirtGenderPipe } from '../../filters/shirt-filter.pipe';
 import { Shirt } from '../../shared/shirt';
 import { ShirtService } from '../../core/shirt.service';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 import { ShoppingCartService } from '../../core/shopping-cart.service';
 import { SlidingPanelsService } from '../../core/sliding-panels.service';
 import { NgbTabset } from '@ng-bootstrap/ng-bootstrap';
